requirements.txt
langgraph>=0.1.0
langchain-anthropic>=0.1.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
plotly>=5.15.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
